* changed PubSub namespace to include OMF version number. IMPORTANT: you need to recreate your pubsub nodes after this patch!! * EC's reset function will no longer send a RESET message to the RC, which is not necessary as it gets rebooted anyway * When the RC gets a second ENROL for the same experiment ID, it will not report this as an error

+++ omf-expctl/ruby/omf-expctl/oml/oml_outputter.rb
@@ -31,7 +31,7 @@ module OMF::EC::OML
 #    stream :logging
 
     param :logger
-    param :level, :type => :long
+    param :level, :type => :int32
     param :level_name
     param :data
     param :tracer
